it's a gun nothing crazy
Basically, I wrote a C++ program called Gun and Bun. It uses OpenCV to watch the webcam. If it sees a person (the 'Bun'), it locks onto them. If it decides you're a threat, it pretends to shoot (the 'Gun'). Just a fun way to mess around with computer vision and automation.